remove duplicate and merge
